+++ tests/integration/test_spansv2.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timezone
+from unittest import mock
 
 from sentry_sdk.envelope import Envelope, Item, PayloadRef
 from sentry_relay.consts import DataCategory
@@ -41,8 +42,18 @@ def envelope_with_spans(*payloads: dict, trace_info=None) -> Envelope:
     return envelope
 
 
-def test_spansv2_basic(spans_consumer, mini_sentry, relay, relay_with_processing):
+def test_spansv2_basic(
+    mini_sentry,
+    relay,
+    relay_with_processing,
+    spans_consumer,
+    metrics_consumer,
+):
+    """
+    A basic test making sure spans can be ingested and have basic normalizations applied.
+    """
     spans_consumer = spans_consumer()
+    metrics_consumer = metrics_consumer()
 
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
@@ -57,7 +68,7 @@ def test_spansv2_basic(spans_consumer, mini_sentry, relay, relay_with_processing
     envelope = envelope_with_spans(
         {
             "start_timestamp": ts.timestamp(),
-            "end_timestamp": ts.timestamp() + 0.500,
+            "end_timestamp": ts.timestamp() + 0.5,
             "trace_id": "5b8efff798038103d269b633813fc60c",
             "span_id": "eee19b7ec3c1b175",
             "name": "some op",
@@ -75,7 +86,7 @@ def test_spansv2_basic(spans_consumer, mini_sentry, relay, relay_with_processing
         "received": time_within(ts),
         "start_timestamp_ms": time_within(ts, precision="ms", expect_resolution="ms"),
         "start_timestamp_precise": time_within(ts),
-        "end_timestamp_precise": time_within(ts),
+        "end_timestamp_precise": time_within(ts.timestamp() + 0.5),
         "duration_ms": 500,
         "exclusive_time_ms": 500.0,
         "is_remote": False,
@@ -87,12 +98,41 @@ def test_spansv2_basic(spans_consumer, mini_sentry, relay, relay_with_processing
         "project_id": 42,
     }
 
+    assert metrics_consumer.get_metrics(n=2, with_headers=False) == [
+        {
+            "name": "c:spans/count_per_root_project@none",
+            "org_id": 1,
+            "project_id": 42,
+            "received_at": time_within_delta(),
+            "retention_days": 90,
+            "tags": {"decision": "keep", "target_project_id": "42"},
+            "timestamp": time_within_delta(),
+            "type": "c",
+            "value": 1.0,
+        },
+        {
+            "name": "c:spans/usage@none",
+            "org_id": 1,
+            "project_id": 42,
+            "received_at": time_within_delta(),
+            "retention_days": 90,
+            "tags": {},
+            "timestamp": time_within_delta(),
+            "type": "c",
+            "value": 1.0,
+        },
+    ]
+
 
 @pytest.mark.parametrize(
     "rule_type",
     ["transaction", "trace"],
 )
 def test_spansv2_ds_drop(mini_sentry, relay, rule_type):
+    """
+    The test asserts that dynamic sampling correctly drops items, based on different rule types
+    and makes sure the correct outcomes and metrics are emitted.
+    """
     project_id = 42
     project_config = mini_sentry.add_full_project_config(project_id)
     project_config["config"]["features"] = [
@@ -107,7 +147,7 @@ def test_spansv2_ds_drop(mini_sentry, relay, rule_type):
     envelope = envelope_with_spans(
         {
             "start_timestamp": ts.timestamp(),
-            "end_timestamp": ts.timestamp() + 0.500,
+            "end_timestamp": ts.timestamp() + 0.5,
             "trace_id": "5b8efff798038103d269b633813fc60c",
             "span_id": "eee19b7ec3c1b175",
             "name": "some op",
@@ -133,7 +173,218 @@ def test_spansv2_ds_drop(mini_sentry, relay, rule_type):
             "timestamp": time_within_delta(),
         },
     ]
-    # TODO: assert span metrics when implemented (usage and count per root)
+
+    assert mini_sentry.get_metrics() == [
+        {
+            "metadata": mock.ANY,
+            "name": "c:spans/count_per_root_project@none",
+            "tags": {"decision": "drop", "target_project_id": "42"},
+            "timestamp": time_within_delta(),
+            "type": "c",
+            "value": 1.0,
+            "width": 1,
+        },
+        {
+            "metadata": mock.ANY,
+            "name": "c:spans/usage@none",
+            "timestamp": time_within_delta(),
+            "type": "c",
+            "value": 1.0,
+            "width": 1,
+        },
+    ]
 
     assert mini_sentry.captured_events.empty()
     assert mini_sentry.captured_outcomes.empty()
+
+
+def test_spansv2_ds_sampled(
+    mini_sentry,
+    relay,
+    relay_with_processing,
+    outcomes_consumer,
+    spans_consumer,
+    metrics_consumer,
+):
+    """
+    The test asserts, that dynamic sampling correctly samples spans with the trace rules
+    of the trace root's project.
+    """
+    outcomes_consumer = outcomes_consumer()
+    spans_consumer = spans_consumer()
+    metrics_consumer = metrics_consumer()
+
+    project_id = 42
+    project_config = mini_sentry.add_full_project_config(project_id)
+    project_config["config"]["features"] = [
+        "organizations:standalone-span-ingestion",
+        "projects:span-v2-experimental-processing",
+    ]
+    _add_sampling_config(project_config, sample_rate=0.0, rule_type="trace")
+
+    sampling_project_id = 43
+    sampling_config = mini_sentry.add_basic_project_config(sampling_project_id)
+    sampling_config["organizationId"] = project_config["organizationId"]
+    _add_sampling_config(sampling_config, sample_rate=0.9, rule_type="trace")
+
+    relay = relay(relay_with_processing(options=TEST_CONFIG), options=TEST_CONFIG)
+
+    ts = datetime.now(timezone.utc)
+    envelope = envelope_with_spans(
+        {
+            "start_timestamp": ts.timestamp(),
+            "end_timestamp": ts.timestamp() + 0.5,
+            "trace_id": "5b8efff798038103d269b633813fc60c",
+            "span_id": "eee19b7ec3c1b175",
+            "name": "some op",
+            "attributes": {"foo": {"value": "bar", "type": "string"}},
+        },
+        trace_info={
+            "trace_id": "5b8efff798038103d269b633813fc60c",
+            "public_key": sampling_config["publicKeys"][0]["publicKey"],
+        },
+    )
+
+    relay.send_envelope(project_id, envelope)
+
+    assert spans_consumer.get_span() == {
+        "trace_id": "5b8efff798038103d269b633813fc60c",
+        "span_id": "eee19b7ec3c1b175",
+        "description": "some op",
+        "data": {
+            "foo": "bar",
+            "sentry.name": "some op",
+            "sentry.server_sample_rate": 0.9,
+        },
+        "measurements": {"server_sample_rate": {"value": 0.9}},
+        "server_sample_rate": 0.9,
+        "received": time_within_delta(ts),
+        "start_timestamp_ms": time_within(ts, precision="ms", expect_resolution="ms"),
+        "start_timestamp_precise": time_within(ts),
+        "end_timestamp_precise": time_within(ts.timestamp() + 0.5),
+        "duration_ms": 500,
+        "exclusive_time_ms": 500.0,
+        "is_remote": False,
+        "is_segment": False,
+        "retention_days": 90,
+        "downsampled_retention_days": 90,
+        "key_id": 123,
+        "organization_id": 1,
+        "project_id": 42,
+    }
+
+    assert metrics_consumer.get_metrics(n=2, with_headers=False) == [
+        {
+            "name": "c:spans/count_per_root_project@none",
+            "org_id": 1,
+            "project_id": 43,
+            "received_at": time_within_delta(),
+            "retention_days": 90,
+            "tags": {"decision": "keep", "target_project_id": "42"},
+            "timestamp": time_within_delta(),
+            "type": "c",
+            "value": 1.0,
+        },
+        {
+            "name": "c:spans/usage@none",
+            "org_id": 1,
+            "project_id": 42,
+            "received_at": time_within_delta(),
+            "retention_days": 90,
+            "tags": {},
+            "timestamp": time_within_delta(),
+            "type": "c",
+            "value": 1.0,
+        },
+    ]
+
+    outcomes_consumer.assert_empty()
+
+
+def test_spansv2_ds_root_in_different_org(
+    mini_sentry,
+    relay,
+    relay_with_processing,
+    outcomes_consumer,
+    spans_consumer,
+    metrics_consumer,
+):
+    """
+    The test asserts that traces where the root originates from a different Sentry organization,
+    correctly uses the dynamic sampling rules of the current project and emits the count_per_root metric
+    into the current project.
+    """
+    outcomes_consumer = outcomes_consumer()
+    spans_consumer = spans_consumer()
+    metrics_consumer = metrics_consumer()
+
+    project_id = 42
+    project_config = mini_sentry.add_full_project_config(project_id)
+    project_config["config"]["features"] = [
+        "organizations:standalone-span-ingestion",
+        "projects:span-v2-experimental-processing",
+    ]
+    _add_sampling_config(project_config, sample_rate=0.0, rule_type="trace")
+
+    sampling_project_id = 43
+    sampling_config = mini_sentry.add_basic_project_config(sampling_project_id)
+    sampling_config["organizationId"] = 99
+    _add_sampling_config(sampling_config, sample_rate=1.0, rule_type="trace")
+
+    relay = relay(relay_with_processing(options=TEST_CONFIG), options=TEST_CONFIG)
+
+    ts = datetime.now(timezone.utc)
+    envelope = envelope_with_spans(
+        {
+            "start_timestamp": ts.timestamp(),
+            "end_timestamp": ts.timestamp() + 0.5,
+            "trace_id": "5b8efff798038103d269b633813fc60c",
+            "span_id": "eee19b7ec3c1b175",
+            "name": "some op",
+            "attributes": {"foo": {"value": "bar", "type": "string"}},
+        },
+        trace_info={
+            "trace_id": "5b8efff798038103d269b633813fc60c",
+            "public_key": sampling_config["publicKeys"][0]["publicKey"],
+        },
+    )
+
+    relay.send_envelope(project_id, envelope)
+
+    assert metrics_consumer.get_metrics(n=2, with_headers=False) == [
+        {
+            "name": "c:spans/count_per_root_project@none",
+            "org_id": 1,
+            "project_id": 42,
+            "received_at": time_within_delta(),
+            "retention_days": 90,
+            "tags": {"decision": "drop", "target_project_id": "42"},
+            "timestamp": time_within_delta(),
+            "type": "c",
+            "value": 1.0,
+        },
+        {
+            "name": "c:spans/usage@none",
+            "org_id": 1,
+            "project_id": 42,
+            "received_at": time_within_delta(),
+            "retention_days": 90,
+            "tags": {},
+            "timestamp": time_within_delta(),
+            "type": "c",
+            "value": 1.0,
+        },
+    ]
+
+    assert outcomes_consumer.get_outcome() == {
+        "category": DataCategory.SPAN_INDEXED.value,
+        "key_id": 123,
+        "org_id": 1,
+        "outcome": 1,
+        "project_id": 42,
+        "quantity": 1,
+        "reason": "Sampled:0",
+        "timestamp": time_within_delta(),
+    }
+
+    spans_consumer.assert_empty()