+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
-**Breaking Changes**:
+**Internal**:
 
-- Only check for local project configs in static mode. ([#5057](https://github.com/getsentry/relay/pull/5057))
+- Set the default log retention to 30 days for full fidelity and downsampled data. ([#5065](https://github.com/getsentry/relay/pull/5065))
+- Improved PII Scrubbing for attributes / logs ([#5061](https://github.com/getsentry/relay/pull/5061)))
+
+## 25.8.0
+
+- Add data categories for Prevent. ([#5052](https://github.com/getsentry/relay/pull/5052))
 
 **Features**:
 
@@ -39,6 +44,7 @@
 - Add producer_name tag, and more rdkafka stats. ([#5031](https://github.com/getsentry/relay/pull/5031))
 - Add `healthcheck` command to relay CLI, for self hosted. ([#5044](https://github.com/getsentry/relay/pull/5044))
 - Change the default configuration to produce to `snuba-items`. ([#5055](https://github.com/getsentry/relay/pull/5055))
+- Backfill client and server sample rates on spans. ([#5059](https://github.com/getsentry/relay/pull/5059))
 
 ## 25.7.0